requirements.txt
tqdm~=2.2.3
psycopg2~=2.8.4
SQLAlchemy~=1.3.11
python-decouple~=3.3
